etl/requirements.txt
pymongo==4.9.2
requests==2.32.3
python-dotenv==1.0.1
dnspython==2.6.1